refactor react - add some typescript types - separate files into manageble chunks with better naming
