+++ frontend/package.json
@@ -32,7 +32,8 @@
     "@babel/register": "^7.0.0",
     "@types/react": "^16.7.13",
     "@types/react-dom": "^16.0.11",
-    "@types/slate": "^0.43.4",
+    "@types/slate": "^0.43.6",
+    "@types/slate-react": "^0.20.3",
     "babel-loader": "^8.0.4",
     "css-loader": "^2.0.0",
     "less-loader": "^4.1.0",
